Copy method getObjectAsJSON to CommandExecutor for be used in enterprise #TASK-5482
