+++ opencga-app/src/main/java/org/opencb/opencga/app/cli/main/executors/OpencgaCommandExecutor.java
@@ -225,90 +225,7 @@ public abstract class OpencgaCommandExecutor extends CommandExecutor {
     }
 
     public String getObjectAsJSON(String objectCategory, String objectPath) throws Exception {
-        StringBuilder jsonInString = new StringBuilder("\n");
-        try {
-            ObjectMap queryParams = new ObjectMap();
-            queryParams.putIfNotEmpty("category", objectCategory);
-            RestResponse<List> response = openCGAClient.getMetaClient().api(queryParams);
-            ObjectMapper jsonObjectMapper = new ObjectMapper();
-           for (List list : response.getResponses().get(0).getResults()) {
-                List<RestCategory> categories = jsonObjectMapper.convertValue(list, new TypeReference<List<RestCategory>>() {});
-                for (RestCategory category : categories) {
-                    for (RestEndpoint endpoint : category.getEndpoints()) {
-                        if (objectPath.equals(endpoint.getPath())) {
-                            boolean enc = false;
-                            for (RestParameter parameter : endpoint.getParameters()) {
-                                //jsonInString += parameter.getName()+":"+parameter.getAllowedValues()+"\n";
-                                if (parameter.getData() != null) {
-                                    enc = true;
-                                    jsonInString.append(printBody(parameter.getData(), ""));
-                                }
-                            }
-                            if (!enc) {
-                                jsonInString.append("No model available");
-                            }
-                            //
-                        }
-                    }
-                }
-           }
-        } catch (Exception e) {
-            jsonInString = new StringBuilder("Data model not found.");
-            CommandLineUtils.error(e);
-        }
-        return jsonInString.toString();
-    }
-
-    private String printBody(List<RestParameter> data, String tabs) {
-        String res = "";
-        res += "{\n";
-        String tab = "    " + tabs;
-        for (RestParameter parameter : data) {
-            if (parameter.getData() == null) {
-                res += printParameter(parameter, tab);
-            } else {
-                res += tab + parameter.getName() + "\"" + ": [" + printBody(parameter.getData(), tab) + "],\n";
-            }
-        }
-        res += tabs + "}";
-        return res;
-
-    }
-
-    private String printParameter(RestParameter parameter, String tab) {
-
-        return tab + "\"" + parameter.getName() + "\"" + ":" + printParameterValue(parameter) + ",\n";
-    }
-
-    private String printParameterValue(RestParameter parameter) {
-
-        if(!StringUtils.isEmpty(parameter.getAllowedValues())){
-            return parameter.getAllowedValues().replace(" ", "|");
-        }
-        switch (parameter.getType()) {
-            case "Boolean":
-            case "java.lang.Boolean":
-                return "false";
-            case "Long":
-            case "Float":
-            case "Double":
-            case "Integer":
-            case "int":
-            case "double":
-            case "float":
-            case "long":
-                return "0";
-            case "List":
-                return "[\"\"]";
-            case "Date":
-                return "\"dd/mm/yyyy\"";
-            case "Map":
-                return "{\"key\": \"value\"}";
-            case "String":
-                return "\"\"";
-            default:
-                return "\"-\"";
-        }
+        return super.getObjectAsJSON(objectCategory, objectPath, openCGAClient);
     }
 
     private boolean isNumeric(String type) {
